+++ src/Chronicle/Configuration/ILoggingConfiguration.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace Chronicle.Configuration
+{
+    interface ILoggingConfiguration
+    {
+        Func<Type, ILogger> LoggerProvider { get; }
+    }
+}