+++ src/Chronicle/Configuration/ILoggingConfigurationContext.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace Chronicle.Configuration
+{
+    public interface ILoggingConfigurationContext
+    {
+        ILoggingConfigurationContext SetLoggerProvider(Func<Type, ILogger> loggerProvider);
+    }
+}