+++ src/Chronicle/Context/LogContext.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace Chronicle.Context
+{
+    static class LogContext
+    {
+        public static Func<Type, ILogger> LoggerProvider { get; set; } = category => new NullLogger();
+    }
+}