+++ CHANGELOG.md
@@ -1,6 +1,38 @@
-# Change Log
+# Changelog
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
+
+### [0.1.1](https://github.com/derekgreer/chronicle/compare/v0.0.3...v0.1.1) (2020-12-04)
+
+
+### Features
+
+* Add support for Serilog ([28556ad](https://github.com/derekgreer/chronicle/commit/28556ad390cc8d877c34117a3895e83e83727ddb))
+* Remove scope use from OperationsTimer. ([81b8479](https://github.com/derekgreer/chronicle/commit/81b847931625672e79f28d41c97d899f48856c17))
+
+### [0.1.2](https://github.com/derekgreer/chronicle/compare/v0.0.3...v0.1.2) (2020-12-04)
+
+
+### Features
+
+* Add support for Serilog ([cdf0cfa](https://github.com/derekgreer/chronicle/commit/cdf0cfaf0f9d269860fbebd26b3f9a4abb6a33f1))
+* Remove scope use from OperationsTimer. ([81b8479](https://github.com/derekgreer/chronicle/commit/81b847931625672e79f28d41c97d899f48856c17))
+
+### [0.1.1](https://github.com/derekgreer/chronicle/compare/v0.0.3...v0.1.1) (2020-12-04)
+
+
+### Features
+
+* Add support for Serilog ([41992df](https://github.com/derekgreer/chronicle/commit/41992df19e179f32729e232a1697e03ccc0e5d64))
+* Remove scope use from OperationsTimer. ([81b8479](https://github.com/derekgreer/chronicle/commit/81b847931625672e79f28d41c97d899f48856c17))
+
+### [0.1.1](https://github.com/derekgreer/chronicle/compare/v0.0.3...v0.1.1) (2020-12-04)
+
+
+### Features
+
+* Add support for Serilog ([5b35006](https://github.com/derekgreer/chronicle/commit/5b35006e32f5e70edd942fb3b1d548905ae216a9))
+* Remove scope use from OperationsTimer. ([81b8479](https://github.com/derekgreer/chronicle/commit/81b847931625672e79f28d41c97d899f48856c17))
 
 <a name="0.1.0"></a>
 # [0.1.0](https://github.com/derekgreer/chronicle/compare/v0.0.3...v0.1.0) (2018-11-27)